Refactor and implement high order component utils for topbar for notification data
